sendind for the final review
